+++ packages/embedded-client/CHANGELOG.md
@@ -14,6 +14,8 @@
 
 - Bump dependency on [@bufbuild/protovalidate] to 1.1.1 ([#1355](https://github.com/cerbos/cerbos-sdk-javascript/pull/1355))
 
+- Bump dependency on [@cerbos/api] to 0.6.0 ([#1358](https://github.com/cerbos/cerbos-sdk-javascript/pull/1358))
+
 ## [0.4.0] - 2026-01-12
 
 ### Added